Remove unnecessary cast on current_temperature. Result from function already a float.

+++ esphome/components/midea_xye/air_conditioner.cpp
@@ -268,7 +268,7 @@ void AirConditioner::ParseResponse(uint8_t cmdSent) {
 
         bool need_publish = false;
 
-        update_property(this->current_temperature, (float) CalculateTemp(RXData[RX_C0_BYTE_T1_TEMP]), need_publish);
+        update_property(this->current_temperature, CalculateTemp(RXData[RX_C0_BYTE_T1_TEMP]), need_publish);
         update_property(this->mode, mode, need_publish);
         if (mode != ClimateMode::CLIMATE_MODE_OFF)  // Don't update below states
                                                     // unless mode is an ON state
